11.12 huge update
bonewall: Made it like a hallucination - purple for aliens
added AlienWalkThrough physicsgroup
bonewall: made it AlienWalkThrough - so aliens can walk through
marine onos stun : lowered 2 second cooldown to 1.5 (need to experiment
with this)
Bonewall: Replaced onupdate with addtimedcallback
bonewall: added check to make sure the model bones are accurate-so no
jitter movement?
sentry: fixed/added stun to sentry - On stun spawns bonewall and
confuses the sentry
so although the onos does 50% less to the sentry (100) - it stuns the
sentry/spawns bonewall on it
and i think confusion effects the fire of the sentry somehow
increased etherealgate lifespan 6 to 12
increased etherealgate cooldown from 8 to 16
increased etherealgate range from 4 to 5
increased etherealgate cost from 5 to 8
known issue: onos can walk through sentry??
fixed aliens 3 levels of health?

untested fix for broken marine animations again
marked bonewall as something to not count as a melee attack to block
against as an alien
meaning, that if you hit someone hitside of a bone well, you will not
hit the bonewall, but instead
the object inside of that bonewall. Where as if the marine where to
shoot at the bonewall, It will
register.
added a way to know when the random passive and automatic research times
for arms lab are set to finish

added 8 second cooldown on scan
corrected mac autospawn - it was spawning 12 - I want to try lowering
that to 8.

macs:
made it so the macs scan the entire map when the front doors are closed,
so that the commander doesnt have to manually click a mac and send it to
200 different locations every damn time something is dropped in a room

wrote a formula to calculate where the "main battle" is happening
and tell every room that is not where this "main battle" is taking
place,
to recieve 2x damage on unpowered structures (only when front doors open
and not siege doors)
(otherwise any room that is not siege room will recieve 2x damage on
structures without power when siege is open)

disabled/removed modifications on power nodes hp/weldrate/buidlrate

main battle location check formula checks every 30 seconds

wrote for armories that are stunned to turn pink/purple just like the
bonewall and becoe
walk through-able and not get in the way of alien melee attacks

added kDistressBeaconCooldown = 10

told the main room to give marine orders to it
told the main room to make a alien threat marker in it

added stunmixin to robofactory and the same settings as armor where to
walk through it when
its stunned with purple effect that also doesnt block alien melee attack

increased marine stun bonewall health 60 to 160
added crag hive upgrade choice #2 (doesnt interferre with choice #1 of
redemption etc)
wrote thickened skin alien upgrade +10% hp - crag hive upgrade choice#2
wrote hunger - +10% HP and +10 energy on player kill - crag hive upgrade
choice#2
wrote for nearby drifters of main room to enzyme the main room if
eligable
wrote for etherealgate to spawn at mainbattle on eligable contestents

replaced spawning etherealgate for main room for turning off lights in
main room
(thanks dark)