1.25
Location: If pregame, and setup, then when teamnumber 1 enters room
during setup,
automatically build node.

GameRules:
When the front doors open, power nodes are counted and defined by two
categories:
Built and not built. (with the exception of siege rooms not being
counted).
The ratio of the two numbers of unbuilt and built are saved for later
use.

When Siege doors open, the ratio count is recounted and compared with
the setup numbers.

From there, after siege opens, every time a main room is set, the ratios
are checked
to determine who owns more map control during siege when compared to
front doors opening.

Right now we're starting off with a simple chance for contamination
based on this number.
Meaning: In this build: if marines hold their ground until siege, then
the more ground
they have defended from setup to siege, increases the chances of a
random contamination spawn
on the powernode.
But, if marines lose their ground, they wont be punished.
However, if marines push their ground and expand territory
that is more than needed, or more than what they had during setup,
then they will be punished. Right now just simply
by increasing the odds of a contamination spawn on the powernode of the
main room.

If the aliens dominate and just take out room after room,
then marines should get an artificial flavoring as well. But this build
does not include that
it was hard enough writing everything as is ;)

Hopefully this may promote both strategies: Teams that choose to defend
the entire map,
and teams that choose to defend purely a room or two. The start will be
to branch
only to buffing during Siege Artificially to see how it goes.

I'm aware this may be flawed in a team may not be setup by setup time.
But im convinced
that this wont be the case unless purposefully exploited.. or low pop
games :/
Either way this is only contamination to start slow and see the chances,
the ifs whens
and how often this actually triggers with the log detailing the ratios.

Id like to add a layer on the minimap with the ratios of unbuilt/built
and say something like
Marine Map Control: x%
Alien Map Control: x%

so the game can get more interactive and dynamic

(Keep in mind the log is outputting all sorts of data to analyze)

Armory/Observatory/PrototypeLab:
Removed alot of research variables and such now that research is no
longer
associated with these, other than Recycle.

PowerPoint:
Set default light mode to nopower

Hive:
Added kHiveMoveUpVector
This way on ns2_epicsiege I may move up hives... ;)

This is it for now until atleast 2pm eastern thanks

p.s ns2_epicsiege expected 1.27