1.28
when a power node dies. From a random duration between 4 and 8 seconds
afterwords,
sends a team notification saying Siege Timer Adjusted.
From there the powerpoint killed/and or built use the same formula.
With the kill multiplying the number by -1 to subtract.
The formula is as follows.
When the power node kills/gets built,
the game rules searches for the closest front door nearest the power
ndoe and gets the distance.
Each distance measurement of 1 is equal to 4 seconds with a minimum sum
of 10 and a maximum of 30
seconds for each current built power node.

**Special rules for 2nd last power node (Aka last room before marine
start)
From there The current time left until siege
is multiplied by 0.35 (so 30 percent of time left until siege)
and that sum is divided by the amount
of power nodes built during setup(-1) (minus one because im not counting
marine start?)
from there that number percentage is diided by
GetReversedRoundLengthToSiege() (Basically
a value between 0 and 1 that says how close the round is to front door
opening compared to siege
so this way a greater bonus is applied the further away from siege time)

So In conclusion the current removal of time is only based on distance
to front door.
Front door room deducting the least amount of time, and the furthest
room away from front.
Aka Arizona Bay, Long Hall to Jetpack Room, the 3rd most outer room
average? Untested on
other maps.

So far the average from memory goes like this on a map with 18 mintimer

13 min front opening
13 to 11
11 to 10/9
9 to 8/7
7 to 4

or if marines only build 1 power node outside of base
the timer bodes from 11 to 4. And if marines build it, goes from 4 to 9.

The max amount never goes aboe the max set at map load and the max
current is 20 min.